Github footer now directs to this website's git
The footer button for github was directing to /solana-labs. I'm assuming
that was a legacy from when labs owned the site and client. Now that
this site is owned by a different repo, I think it should be pointed
there so contributors know where to go.

There's also an argument for a github link to direct to a protocol
client repo, but that should live somewhere else because that's 3
different repos at this point.